Revert "Squashed commit of the following:"
This reverts commit 6341baa0619ac47d5153ba30c4ee7e4e26e9da63.

+++ Mythic Chronicles Sentinel/Assets/Scripts/HealthBar.cs
@@ -7,12 +7,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider healthSlider;
 
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         healthSlider.value = health;
     }
 
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         healthSlider.maxValue = health;
         healthSlider.value = health;
